feat: Credentials Authentication This PR adds email / password authentication to create-t3-turbo using the NextAuth Credentials Provider.
Although NextJS discourages password authentication, for companies with an existing user base with existing credentials, it is difficult to not support password authentication. NextAuth also attempts to disable database sessions if using the Credentials Provider.

This PR demonstrates how to use the NextAuth Credential Provider to create JWT tokens and Database sessions.

The `authorize` method validates the email, hashes the password with brypt, compares password hashes, creates a database session, and returns the user object and sessionToken.

The `jwt` callback is used to add the sessionToken to the token object used to generate the JWT, tying the JWT to the database session, and validates the current JWT token against the database.

A `signOut` event is used delete the session from the database if the sessionToken is present in the JWT.

This should work with additional providers such as discord.